embedding-api/requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ColPali and ML dependencies
colpali-engine>=0.1.0
torch>=2.0.0
transformers>=4.35.0
Pillow>=9.5.0
numpy>=1.24.0

# HTTP client (for testing)
httpx>=0.25.0

# Optimization
flash-attn>=2.3.0; platform_machine=="x86_64" and platform_system=="Linux"